+++ js/embedded/embeddedplayer.js
@@ -322,6 +322,12 @@ OCA.Music.EmbeddedPlayer = function() {
 		parentContainer.resize(resizeControls);
 		resizeControls();
 
+		// While the z-index from CSS is crucial on ownCloud and older Nextcloud versions, NC28+ doesn't need it. In addition, there it
+		// causes the "overlay" scrollbar used by Firefox on Windows 11 to be hidden behind the pane which we don't want.
+		if (OCA.Music.Utils.getScrollContainer().is($('#app-content-vue .files-list'))) {
+			musicControls.css('z-index', 'unset');
+		}
+
 		// bind the player events to change the states of the UI controls
 		player.on('play', () => {
 			playButton.css('display', 'none');